Do not emit the parentheses wrapping the default value with `#[default(...)]`
This fixes the compiler suggestions; see rust-lang/rust#94074.

Technically this is a breaking change - previously, because of the parentheses, you could actually pass a tuple as the default value without wrapping it in parentheses - e.g. `#[default(1, 1)]`. Now you need to do `#[default((1, 1))]`. However, I don't think someone relied on this, as it looks like the wrong behavior.